refactor(engine): split wiring.rs into responsibility submodules
The 827-line primer-engine/src/wiring.rs becomes a directory-style
module split by responsibility. Behaviour-preserving: identical
14-symbol external pub surface, so every primer_engine::wiring::<name>
path (and the crate-root re-exports in lib.rs) is unchanged.

All files now under 500 lines:
- wiring/mod.rs (184) — shared BackendParams input struct + the QAIRT
  path helpers (default_qairt_lib_dir / resolve_qairt_lib_dir /
  resolve_qairt_lib_dir_for) + the flat `pub use` re-export facade + the
  seven already-extracted test-module declarations. These stay here so
  the submodules and qnn_dispatch_tests borrow them via `super::`.
- wiring/backend.rs (337) — the `--backend` dispatch matrix, MainBackendPlan
  / plan_main_backend / resolve_fallback_model, build_main_backend's opt-in
  fallback, the RouterBackend path, and the feature-gated qnn/llamacpp
  constructors (the qnn arm calls `super::resolve_qairt_lib_dir`).
- wiring/subsystems.rs (241) — build_classifier / build_extractor /
  build_comprehension.
- wiring/embedders.rs (128) — the three Embedder backends.

Four `wiring/*_tests.rs` files gain explicit `use` lines for aliases the
old flat `use` block leaked into their `use super::*`
(ClassifierSettings / ExtractorSettings / InferenceBackend / Arc).

Verification:
- pub-surface diff old-vs-new: empty (identical 14 symbols).
- cargo test -p primer-engine: 72 passed / 0 failed (baseline match).
- cargo clippy -p primer-engine --all-targets -- -D warnings: clean.
- feature-combos clippy (qnn,llamacpp,embedding,ollama-embedding,
  openai-compat-embedding) -- -D warnings: clean (exercises every
  feature-on arm in backend.rs + embedders.rs).
- cargo fmt --check: clean.
- full-workspace cargo test: 51 test-result-ok, 0 failed.

README/ROADMAP untouched (internal file split changes neither
user-facing status nor the phase plan). CLAUDE.md gets a module-layout
note on the primer-engine bullet.
